Update Logic Puzzle - NEEDS CHANGES
Didn't connect the nodes, and created a glitch that doubled the nodes when above them - also couldn't win. UI was messed up, created a new connection instead of actually connecting to the node previously selected, and didn't reset when I messed up. Also, connection of nodes needed to be line based and some nodes unused. Got so glitchy it was unusable.